Deploy/requirements.txt
pandas
numpy
streamlit
streamlit-chat
requests
joblib
openai
keras
tensorflow
python-dotenv
scikit-learn
